# gen/http/openapi.yaml
swagger: "2.0"
info:
    title: starterpack
    description: Best API REST building with GoaDesign
    contact:
        name: Guillaume MORIN
        email: guillaume.morin@epitech.eu
        url: https://guillaume-morin.fr/
    license:
        name: GOA X GM
        url: https://guillaume-morin.fr/
    version: "1.0"
host: localhost:8088
consumes:
    - application/json
    - application/xml
    - application/gob
produces:
    - application/json
    - application/xml
    - application/gob
paths:
...
    /refresh:
        post:
            tags:
                - jwtToken
            summary: refresh jwtToken
            description: Refresh Token
            operationId: jwtToken#refresh
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: RefreshRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/JWTTokenRefreshRequestBody'
                    required:
                        - refresh_token
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/JWTTokenRefreshResponseBody'
                        required:
                            - access_token
                            - refresh_token
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/JWTTokenRefreshUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
...
    /v1/file/add:
        post:
            tags:
                - files
            summary: importFile files
            description: Import file
            operationId: files#importFile
            consumes:
                - multipart/form-data
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: jwtToken
                  in: header
                  description: Jwt token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: ImportFileRequestBody
                  in: formData
                  required: true
                  schema:
                    $ref: '#/definitions/FilesImportFileRequestBody'
                    required:
                        - filename
                        - content
                        - format
            responses:
                "201":
                    description: Created response.
                    schema:
                        $ref: '#/definitions/FilesImportFileResponseBody'
                        required:
                            - success
                            - file
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/FilesImportFileUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
                  jwt_header_jwtToken: []
    /v1/file/remove:
        patch:
            tags:
                - files
            summary: deleteFile files
            description: Delete one file by ID
            operationId: files#deleteFile
            parameters:
                - name: Authorization
                  in: header
                  description: OAuth token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: jwtToken
                  in: header
                  description: Jwt token
                  required: false
                  type: string
                  pattern: ^Bearer [^ ]+$
                - name: DeleteFileRequestBody
                  in: body
                  required: true
                  schema:
                    $ref: '#/definitions/FilesDeleteFileRequestBody'
                    required:
                        - url
            responses:
                "200":
                    description: OK response.
                    schema:
                        $ref: '#/definitions/FilesDeleteFileResponseBody'
                        required:
                            - success
                "500":
                    description: Internal Server Error response.
                    schema:
                        $ref: '#/definitions/FilesDeleteFileUnknownErrorResponseBody'
                        required:
                            - err
                            - success
                            - error_code
            schemes:
                - http
            security:
                - OAuth2_header_Authorization: []
                  jwt_header_jwtToken: []
...
definitions:
...
    FilesDeleteFileRequestBody:
        title: FilesDeleteFileRequestBody
        type: object
        properties:
            url:
                type: string
                example: /public/uploads/2021/12/2ca51d10-b660-4b2c-b27f-f7a119642885.png
                minLength: 23
        example:
            url: /public/uploads/2021/12/2ca51d10-b660-4b2c-b27f-f7a119642885.png
        required:
            - url
    FilesDeleteFileResponseBody:
        title: FilesDeleteFileResponseBody
        type: object
        properties:
            success:
                type: boolean
                example: false
        example:
            success: false
        required:
            - success
    FilesDeleteFileUnknownErrorResponseBody:
        title: FilesDeleteFileUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: false
        required:
            - err
            - success
            - error_code
    FilesImportFileRequestBody:
        title: FilesImportFileRequestBody
        type: object
        properties:
            content:
                type: string
                description: content of image
                example:
                    - 85
                    - 110
                    - 100
                    - 101
                    - 32
                    - 110
                    - 101
                    - 115
                    - 99
                    - 105
                    - 117
                    - 110
                    - 116
                    - 32
                    - 99
                    - 117
                    - 109
                    - 113
                    - 117
                    - 101
                    - 46
                format: byte
            filename:
                type: string
                description: uploaded file name
                example: foo.jpg
                minLength: 2
            format:
                type: string
                description: uploaded file format
                example: image/jpeg
                enum:
                    - image/jpeg
                    - image/png
                    - image/jpg
            h:
                type: integer
                description: height of image if you crop
                example: 8529679476592417868
                format: int64
            mime:
                type: string
                description: url file
                example: Illo repudiandae in eum ratione sed.
            size:
                type: integer
                description: size of image
                example: 5795732347313123467
                format: int64
            url:
                type: string
                description: url file
                example: Molestias ad quasi sed laborum.
            w:
                type: integer
                description: width of image if you crop
                example: 2908476327083422861
                format: int64
        example:
            content:
                - 81
                - 117
                - 105
                - 32
                - 100
                - 117
                - 99
                - 105
                - 109
                - 117
                - 115
                - 46
            filename: foo.jpg
            format: image/jpeg
            h: 5194974390363827584
            mime: Quia quam et incidunt vero.
            size: 1352988947556285481
            url: Vero deleniti sint temporibus deserunt.
            w: 5381470590522431180
        required:
            - filename
            - content
            - format
    FilesImportFileResponseBody:
        title: FilesImportFileResponseBody
        type: object
        properties:
            file:
                $ref: '#/definitions/resFileResponseBody'
            success:
                type: boolean
                example: true
        example:
            file:
                id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
                mime: Animi quia eos est ipsam magnam.
                name: foo.png
                size: 6063255645832637196
                url: Eum non delectus qui explicabo.
            success: true
        required:
            - success
            - file
    FilesImportFileUnknownErrorResponseBody:
        title: FilesImportFileUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: true
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: true
        required:
            - err
            - success
            - error_code
    JWTTokenRefreshRequestBody:
        title: JWTTokenRefreshRequestBody
        type: object
        properties:
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
                minLength: 32
        example:
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
        required:
            - refresh_token
    JWTTokenRefreshResponseBody:
        title: JWTTokenRefreshResponseBody
        type: object
        properties:
            access_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token:
                type: string
                example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success:
                type: boolean
                example: true
        example:
            access_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            refresh_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiYWRtaW4iOnRydWV9.TJVA95OrM7E2cBab30RMHrHDcEfxjoYZgeFONFh7HgQ
            success: false
        required:
            - access_token
            - refresh_token
            - success
    JWTTokenRefreshUnknownErrorResponseBody:
        title: JWTTokenRefreshUnknownErrorResponseBody
        type: object
        properties:
            err:
                type: string
                example: sql no rows affected
            error_code:
                type: string
                example: TX_UPDATE_ITEM
            success:
                type: boolean
                default: false
                example: false
        description: Error not identified (500)
        example:
            err: sql no rows affected
            error_code: TX_UPDATE_ITEM
            success: true
        required:
            - err
            - success
            - error_code
...
    resFileResponseBody:
        title: resFileResponseBody
        type: object
        properties:
            id:
                type: string
                example: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
                format: uuid
            mime:
                type: string
                example: Tempore aut.
            name:
                type: string
                example: foo.png
            size:
                type: integer
                example: 5628325901883333098
                format: int64
            url:
                type: string
                example: Quo voluptas laboriosam laborum repellat consectetur magnam.
        example:
            id: 5dfb0bf7-597a-4250-b7ad-63a43ff59c25
            mime: Est aut odit aliquid.
            name: foo.png
            size: 1532588874162345822
            url: Et animi et.
        required:
            - id
            - name
            - url
...
securityDefinitions:
    OAuth2_header_Authorization:
        type: oauth2
        description: Use OAuth2 to authenticate
        flow: application
        tokenUrl: /authorization
        scopes:
            api:read: Provides read access
    jwt_header_jwtToken:
        type: apiKey
        description: |-
            Secures endpoint by requiring a valid JWT token retrieved via the signin endpoint. Supports scopes "api:read" and "api:write".

            **Security Scopes**:
              * `api:read`: Read-only access
              * `api:write`: Read and write access
        name: jwtToken
        in: header